### 2016-07-26 #60 @seaaan
plateR: tools for microtiter plate-shaped data
It makes it easy to convert between plate-shaped data and tidy data. [Plates](https://www.google.com/search?q=microtiter+plate&tbm=isch) are used to perform experiments. Scientific instruments often give back data shaped like a plate, which is terrible for analysis. plateR lets you store plate-shaped data (easy to think about) but import tidy data (easy to analyze).

```
Package: plateR
Title: Tools to Make it Easy to Work with Microtiter Plate-Shaped Data
Version: 0.2.1
Authors@R: person("Sean", "Hughes", email = "smhughes@uw.edu",
    role = c("aut", "cre"))
Description: Tools for interacting with data in the form of microtiter plates.
    plateR defines a simple, plate-shaped file format for data storage, so it's
    easy to remember the experimental design, and provides functions to
    seamlessly convert between that format and a tidy data frame that's optimal
    for analysis.
    knitr
    dplyr

```

https://github.com/seaaan/plateR
Lab scientists
Not that I know of

    - [X] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.

The title is not completely in lower-case. I am open to changing it to all lower-case. My rationale for the capital R was to distinguish it from the word plater.

NA

Possibly @daattali
