### 2016-07-26 #60 @sckott
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)

---

#### Editor comments

@seaaan  Thanks for your submission!

One big thing has not bee done

- Examples are needed, e.g., the checking examples line from R check

```
─  checking examples ... NONE
```

Please add examples for all exported functions, we'll continue review once that is done.
