### 2016-07-26 #60 @daattali
I don't mind doing a review, but I'd be able to review the code from a technical point of view much better than practical.  I have extremely limited experience with plates, I've only ever dealt with one kind (ddpcr by bio-rad) so I don't really know the ways this package could be applied.  I can definitely assume that this package provides a useful utility for many people, but I myself don't know enough about plate files in the wild to judge that.  For example, I looked at the [sample input](https://github.com/seaaan/plateR/blob/master/inst/extdata/example-1.csv) and it seems weird to me - is that a format that is commonly seen, or is that what users are expected to manually transform their data into?

So sure, I'll review if you think it's ok for me to do so without the domain specific knowledge :)
