+++ R/view_plate.R
@@ -11,6 +11,17 @@
 #' @return A depiction of the data in \code{column_to_display} as 
 #' though laid out on a microtiter plate with \code{plate_size} wells.
 #' @export
+#' @examples 
+#' # Get some tidy data
+#' file_path <- system.file("extdata", "example-1.csv", package = "plateR")
+#' data <- read_plate(file = file_path, well_ids_column = "Wells")
+#' head(data)
+#' 
+#' # See which wells got which drug
+#' view_plate(data, "Wells", "Drug", 96)
+#' 
+#' # See concentration of the drugs across the plate
+#' view_plate(data, "Wells", "Concentration", 96)
 view_plate <- function(data, well_ids_column, column_to_display, 
                       plate_size = 96) {
    # validate well_ids_column
