+++ R/add_plate.R
@@ -16,6 +16,23 @@
 #' @return Returns data with as many new columns as plates in \code{file}. 
 #' Empty wells are indicated with NA.
 #' @export
+#' @examples 
+#' 
+#' # Part of the data is tidy
+#' file <- system.file("extdata", "example-2-part-A.csv", package = "plateR")
+#' data <- read.csv(file)
+#' 
+#' # Part of the data is plate-shaped
+#' plate_shaped <- system.file("extdata", "example-2-part-B.csv", package = "plateR")
+#' 
+#' # Combine the two
+#' data <- add_plate(
+#'    file = plate_shaped,
+#'    data = data, 
+#'    well_ids_column = "Wells")
+#' 
+#' # Now data are tidy
+#' head(data)
 add_plate <- function(file, data, well_ids_column) {
    to_add <- read_plate(file, "wellIds")
    
