### 2016-07-26 5d1d373
Update vignette to include read_plates
