### 2016-07-26 c55a0e2
Change view_plate example to one that doesn't break devtools::check()
