+++ R/add_plate.R
@@ -17,7 +17,6 @@
 #' Empty wells are indicated with NA.
 #' @export
 #' @examples 
-#' 
 #' # Part of the data is tidy
 #' file <- system.file("extdata", "example-2-part-A.csv", package = "plateR")
 #' data <- read.csv(file)
