### 2016-07-26 #60 @daattali
Adding examples to the functions is a good idea even if they're replicated
from the vignette because that way it's easily accessible when you're
inside R and are looking at the docs of a function and want to see its
usage.

In regards to my question, I was wondering specifically if the format of
having several "tables" inside a single csv is common, or is that what the
users need to create? That's the format that seems strange to me
