### 2016-07-26 #60 @seaaan
Ah, misinterpreted your question. No, the several tables inside a single csv is not common, but it turned out to be convenient (at least for me!).
