### 2016-07-26 #60 @sckott
thx for adding examples @seaaan

I'll assign @daattali and look for another reviewer that has domain knowledge
