### 2016-08-04 #60 @daattali
I see there's a `reviewers-assigned` label, so who are the two reviewers, and what is the deadline?
