### 2016-08-04 #60 @sckott
sorry, edited comment above https://github.com/ropensci/onboarding/issues/60#issuecomment-235349786

thanks @jooolia and @daattali
