### 2016-08-08 #60 @daattali
##### Paper (for packages co-submitting to JOSS)

The package contains a `paper.md` with:

- [x] **A statement of need** clearly staing problems the software is designed to solve and its target audience.

- [ ] **The author has responded to my review and made changes to my satisfaction. I recemmend approving this package.**

---

When I first saw the package I didn't know what it could be used for because I'm not too familiar with plate files and only have experience with a single kind, but after taking a deeper look I can 100% support this package, it can definitely be very useful for scientists dealing with plates with multiple variables.  Overall this package was written, documented, and tested very well and in a very maintainable fashion.

- reminder to include  rOpenSci footer image after the package gets approved
- I suppose there's no harm in README.Rmd but I'll just mention that in this repo it seems unnecessary since there's no R code in your README. You could safely delete README.Rmd and simply use a simple markdown format. But feel free to keep it if you prefer, just wanted to mention this
- README doesn't explicitly say who the target audience is, but I feel like it's fine, it's very trivial to realize who can use it
- There is no Maintainer field in DESCRIPTION, but I think it will get generated automatically from the "Authors@R" field maybe? @sckott maybe this could be clarified in the reviewer guidelines?

- I recommend adding documentation for "plateR" as the package, so that if I just want to get to the index page of the plateR documentation without a specific function in mind, I could just type "?plateR"
- the "file" param in "read_plate" says "A character vector with the path of a..." which gives the impression that you can pass in multiple files (because of the name vector). Is this true? If you only expect one file, then I would change the wording
- The plateR format should be explained elsewhere , not only in the read_plate function doc. The vignette does have a section about the plateR format but it only explains it in words. It would be very useful to see some examples within the vignette (I had to look at the example files under /inst/extdata in order to understand what the file should look like). Perhaps even a simple graphic
- The documentation for plateR format for having multiple columns says "After the first plate, leave one row blank" -- by "blank" do you mean a row of commas with as many commas as there are columns? I thought blank means literally a blank line, but looking at the example csv files I see ",,,,,,,,,,,"

- Not clear why the # character cannot be used inside a plateR file
- How is the user meant to generate the input file, is ot meant to be done by hand? Can there be any tool provided to help with creating this plateR file format?
- It could be useful to have a separate function that just checks a file to see if it is in proper plateR format , since it's a format that needs to be done manually and is non standard
- For `read_plates()`, I think it'd be nicer to have the plate name column as the first column rather than the last , just so that when there are many variables, it's easy to see the plate right away. You call!
- Another "your call" suggestion: it seems to me that as a user `read_plate()` and `read_plates()` are essentially the same thing, except if there are multiple files then there is also the additional optional parameter of `plate_names`. If that's the only difference, it might be a bit simpler to just combine them into one function that takes either one or multiple files, instead of providing two essentially identical functions that only differ in the number of input files
- Is there a good reason why `well_ids_column` has a default value of "Wells" in some functions but not in `add_plate()`? It would be more consistent to just always use that default value
- `add_plate()`: Have you thought about making `data` the first param and `file` the second? Then you could do something like this: `read.csv(file1) %>% add_plate(file2) %>% add_plate(file3)` I'm not 100% sure this interaction makes sense, just wanted to throw it out there
- The `view_plate()` function is really cool! Have you thought about trying to support multiple `column_to_display` and just showing each one on a separate line? Maybe it can be useful if you want to visualize two variables at a time?
- It's really helpful to get an error message when reading a plate that tells me exactly in which layout there was a problem
- Is it really necessary to provide the `plate_size` to `view_plate()`? Since the data is already provided, can't it just figure out the size?
- There's some very good error checking in some places, but in some functions the error checking and  input sanitization could be improve. For example, I shouldn't be allowed to pass `well_ids_column = ""` and I should get a clear error if I pass an invalid string to `file` or a `NULL` value to  `well_ids_column`. I also noticed I get a non descriptive error if an input csv file is empty

#### Source code
- In [guess_plate_size](https://github.com/seaaan/plateR/blob/b204d44c0a852543a0ba011d7d9f307a855f34cf/R/plate_utilities.R#L125): I would think it's safer to use `vector[-1]` instead of `vector[2:length()]` (for example, if the vector only has one element)
- In [get_plate_size_from_number_of_columns](https://github.com/seaaan/plateR/blob/b204d44c0a852543a0ba011d7d9f307a855f34cf/R/plate_utilities.R#L70): I don't think the return type is meant to say that it will return TRUE
- `get_plate_size_from_number_of_columns()`: maybe throwing an error makes more sense than returning a string indicating an error when an invalid number if supplied. If the only reason you returned a string is because you don't want to have an error thrown from within a nested helper function, you could use `tryCatch()` to see if that function resulted in an error instead of checking for the return type
- `plate_dimensions()` To me it looks like random numbers , is that common accepted knowledge that plates only come in those 5 standard sizes?
- [`read_plate()`](https://github.com/seaaan/plateR/blob/b204d44c0a852543a0ba011d7d9f307a855f34cf/R/read_plate.R#L117): it feels a bit hacky to "try adding one in case the file ends in a blank row" -- this method would also succeed is there is an extra row that isn't blank, right?  Could you instead explicitly check if the last row is blank and then you know whether or not to remove it
- I see many uses of mapply and sapply, it might be safer to try to stay away from those functions altogether
- Really good use of helper functions
- I think the only reason you import dplyr is for `bind_row`. If that's indeed the only reason, it might be worth it to just use a base equivalent and remove the dependency on such a heavy package (even though it's safe to assume most people will have it installed). I think something like `do.call(rbind, list_of_df)` will do the job
- This is a very negligible comment but I can't not say it :) the `validate_column_is_in_data()` could take in a vector of columns and then you don't need to call it twice in a row with different columns
- I see a lot of dataframe subsettings and very rarely a use of `drop = FALSE`. Have you verified that all the functions work as intended when the input file has only one well/one row/one column?

- WOW very extensive testing! Well done, very thorough

#### Misc

- I wouldn't put LICENSE in buildignore, I'd want CRAN to know about the license
- in the YAML of paper.md I would remove the indentation that you have for all the fields
- DESCRIPTION file: I would recommend placing the suggests, depends, and imports one after the other so that it's easy to see at a glance all the used packages, rather than having those fields in different parts of the file
- I recommend adding dates to the NEWS file
- the cran-comments file should specify which version of the package is submitted for each iteration of the comments (by looking at the file right now I have no idea what version is on CRAN)
- The second sentence of the package description (in DESCRIPTION) is a big long, perhaps break it into two sentences. I noticed it's the same sentence as in the paper.md but in paper.md it sounds better because it's not the second and last sentence
- The name of the package being "plater" vs "plateR" is a personal preference. I personally prefer all lower case to make it easier to type and so that you don't have to remember what letters are capitalized, but it's up to you (or up to the rOpenSci editors)

I hope my comments make sense and are helpful. These are all of course guidelines and open to discussion, I just brain-dumped as I went through the code
