### 2016-08-16 #60 @ropenscibot
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ ] **Archive** (JOSS only may be post-review): The repository DOI resolves correctly
- [ ] **Version** (JOSS only may be post-review): Does the release version given match the GitHub release (v1.0.0)?

---

#### Editor comments

@seaaan  Thanks for your submission!

One big thing has not bee done

- Examples are needed e.g. the checking examples line from R check

```
─  checking examples ... NONE
```

Please add examples for all exported functions we'll continue review once that is done.

---

 @daattali  @jooolia
Due date: 2016-08-23 - hey there, it's been 21 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
