### 2016-08-16 #60 @daattali
The bot is a bit strict there - @jooolia still has 8 days until the deadline
