### 2016-08-17 #60 @seaaan
Thank you very much for your detailed and thoughtful review. I understand
your concerns about the narrowness of the focus and will think about how to
address that as I make changes in response to both reviews. Thanks again to
both of you!

On Wed, Aug 17, 2016 at 8:06 AM, Julia Gustavsen <notifications@github.com>
wrote:

> Package Review
>
> *Please check off boxes as applicable, and elaborate in comments below.
> Your review is not limited to these topics, as described in the reviewer
> guide*
>
>    - As the reviewer I confirm that there are no conflicts of interest
>    for me to review this work (such as being a major contributor to the
>    software).
>
> Documentation
>
> The package includes all the following forms of documentation:
>
>    - *A statement of need* clearly stating problems the software is
>    designed to solve and its target audience in README
>    - *Installation instructions:* for the development version of package
>    and any non-standard dependencies in README
>    - *Vignette(s)* demonstrating major functionality that runs
>    successfully locally
>    - *Function Documentation:* for all exported functions in R help
>    - *Examples* for all exported functions in R Help that run
>    successfully locally
>    - *Community guidelines* including contribution guidelines in the
>    README or CONTRIBUTING, and URL, Maintainer and BugReports fields in
>    DESCRIPTION
>
> Paper (for packages co-submitting to JOSS)
>
> The package contains a paper.md with:
>
>    - *A short summary* describing the high-level functionality of the
>    software
>    - *Authors:* A list of authors with their affiliations
>    - *A statement of need* clearly staing problems the software is
>    designed to solve and its target audience.
>    - *References:* with DOIs for all those that have one (e.g. papers,
>    datasets, software).
>
> Functionality
>
>    - *Installation:* Installation succeeds as documented.
>    - *Functionality:* Any functional claims of the software been
>    confirmed.
>    - *Performance:* Any performance claims of the software been confirmed.
>    - *Automated tests:* Unit tests cover essential functions of the
>    package and a reasonable range of inputs and conditions. All tests pass on
>    the local machine.
>    - *Packaging guidelines*: The package conforms to the rOpenSci
>    packaging guidelines
>
> Final approval (post-review)
>
>    - *The author has responded to my review and made changes to my
>    satisfaction. I recommend approving this package.*
>
> Estimated hours spent reviewing: 5 hours Review Comments Overall
>
> This package is very well documented and coded. For the package's purpose,
> I found that the tests were very extensive and covered many different
> cases. I had no installation problems under Windows and Linux. Any weird
> data that I could think to send through the package was met with an
> informative and useful error message. The package build, checks, and tests
> all passed on my machine. The functions, variables, and arguments are
> well-named and easy to understand. The vignette gives a nice overview of
> what the package does, and the help documentation and examples are
> well-written. I did not come across anything that I thought could be
> improved with how the package was constructed. I also thought the
> flexibility of plate size (# of wells) and the flexibility of the kinds of
> data (wells can contain numeric or character data) was very well
> thought-out. The only thing I would add would be documentation for plateR
> as a package (so that you can use help("plateR") and get the overall help
> menu).
> Data formats
>
> The following is solely to do with the audience and applicability and is
> just based on my experience. I do not want to seem overly negative, but the
> only downside I see for this package is that I'm not sure how many
> scientists will have data that look like your example data. The example
> data are very clear and you have made it very clear how to recreate the
> files. I realize that often the software used to get the data off of
> different machines in proprietary so it is neither practical nor possible
> to have an R package that would deal with all these different formats (nor
> is that the goal of the package!). I thought I would be able to test out
> this package using qPCR data that I had, but the format of my data was that
> either each well was a row and Cq, Dilution, SQ, etc were different columns
> (so already tidy) or each well is a column and each row is a cycle.
> Therefore to work with the second kind of data in plateR I would have to
> read in the excel file and then manipulate the data before using plateR
> (thus would likely end up with tidy data through another route). I did have
> some spectrophotometric data I received and sequencing plate layouts I
> submitted in a tidy format that I could have transformed back into a plate
> format using plateR, but for this to be really useful I think some sort of
> visualization of the values would be helpful (e.g. looking at the "microtiter"
> section of Bioconductor.org
> <https://www.bioconductor.org/packages/release/BiocViews.html#___MicrotitrePlateAssay>
> see packages that do this like splots
> <https://www.bioconductor.org/packages/release/bioc/html/splots.html> or
> more specific qPCR viz in HTqPCR
> <https://www.bioconductor.org/packages/release/bioc/html/HTqPCR.html>). I
> am not suggesting you need to do any of this, just providing comments based
> on my experience.
> Replicates?
>
> Another thing that is often useful and important when analysing data from
> microtiter plates is being able to easily group together biological and/or
> technical replicates. Since plateR already makes use of dplyr I could
> imagine that either a function or a demonstration using group_by() (or
> another approach) could be helpful for analysing these kinds of data.
> Summary
>
> In summary, this package really excels in its stated purpose of reading in
> data that are in the specific format of looking like a microtiter plate,
> its testing, and its documentation. Just to be clear, I do not require any
> changes for the package to be approved. Thanks for the chance to review!
> Session info
>
> sessionInfo()
>
> ## R version 3.3.0 (2016-05-03)
> ## Platform: x86_64-redhat-linux-gnu (64-bit)
> ## Running under: Fedora 22 (Twenty Two)
> ##
> ## locale:
> ##  [1] LC_CTYPE=en_CA.UTF-8       LC_NUMERIC=C
> ##  [3] LC_TIME=en_CA.UTF-8        LC_COLLATE=en_CA.UTF-8
> ##  [5] LC_MONETARY=en_CA.UTF-8    LC_MESSAGES=en_CA.UTF-8
> ##  [7] LC_PAPER=en_CA.UTF-8       LC_NAME=C
> ##  [9] LC_ADDRESS=C               LC_TELEPHONE=C
> ## [11] LC_MEASUREMENT=en_CA.UTF-8 LC_IDENTIFICATION=C
> ##
> ## attached base packages:
> ## [1] stats     graphics  grDevices utils     datasets  methods   base
> ##
> ## other attached packages:
> ## [1] plateR_0.2.1
> ##
> ## loaded via a namespace (and not attached):
> ##  [1] R6_2.1.2        assertthat_0.1  magrittr_1.5    DBI_0.5
> ##  [5] tools_3.3.0     htmltools_0.3.5 dplyr_0.5.0     tibble_1.1
> ##  [9] yaml_2.1.13     Rcpp_0.12.6     stringi_1.1.1   rmarkdown_1.0
> ## [13] knitr_1.14      stringr_1.0.0   digest_0.6.10   evaluate_0.9
>
