### 2016-08-17 #60 @sckott
Thanks for the reviews @jooolia and @daattali

@seaaan continue the conversation here
