### 2016-08-17 56a7c86
Make file path of read_plate less confusing.
