+++ man/read_plate.Rd
@@ -7,8 +7,7 @@
 read_plate(file, well_ids_column = "Wells")
 }
 \arguments{
-\item{file}{A character vector with the path of a .csv file formatted as 
-described below.}
+\item{file}{The path of a .csv file formatted as described below.}
 
 \item{well_ids_column}{The name to give the column that will contain the well
 identifiers. Default "Wells".}
