### 2016-08-17 39e9c57
Clarify what is meant by a "blank" row between layouts.
