+++ inst/doc/plateR-basics.md
@@ -35,7 +35,7 @@ The format is pretty simple:
 -   Top left cell of each layout is the name
 -   The rest of the top row of each layout is the column numbers (1:12 for a 96-well plate)
 -   The rest of the left column is the row names (A:H for a 96-well plate)
--   One line between layouts
+-   One line between layouts (This row should appear as blank in a spreadsheet editor, but as a row of commas when viewed as plain text.)
 
 You can use `plateR` format with any standard plate size (6 to 384 wells). Not every well has to be filled. If a well is blank in every layout in a file, it's omitted. If it's blank in some but not others, it'll get `NA` where it's blank.
 
