### 2016-08-18 1bdd096
Make Plate the first column in the result of read_plates()
