### 2016-08-19 c8c032c
Reorder arguments of add_plate
