### 2016-08-19 23b85a9
Add rmarkdown to suggests
