### 2016-08-19 50f9b90
Remove test that was only relevant to earlier version of the package.
