### 2016-08-19 81f741d
Safer elimination of first element of vector
