### 2016-08-19 7a23b94
Fix error in documentation of get_plate_size_from_number_of_columns
