### 2016-08-19 b0ea508
Throw error in get_plate_size_from_number_of_columns instead of wrapping it.
