### 2016-08-19 e1156b7
Make validate_column_is_in_data take a vector of columns instead of one column and add tests for it.
