### 2016-08-19 2978589
Remove unnecessary tabs from paper.md
