### 2016-08-24 b5a3082
Put dependencies in order in the DESCRIPTION
