+++ NEWS.md
@@ -1,3 +1,7 @@
+# plateR 0.2.2 (24 Aug 2016)
+* Changes in response to rOpenSci reviewers
+* Reorder arguments of add_plate for better pipelining
+
 # plateR 0.2.1
 * Reorganize parameters for consistency
 * Add defaults for parameters
