+++ DESCRIPTION
@@ -5,7 +5,7 @@ Authors@R: person("Sean", "Hughes", email = "smhughes@uw.edu",
     role = c("aut", "cre"))
 Description: Tools for interacting with data in the form of microtiter plates.
     plateR defines a simple, plate-shaped file format for data storage, so it's 
-    easy to remember the experimental design, and provides functions to 
+    easy to remember the experimental design. The package provides functions to 
     seamlessly convert between that format and a tidy data frame that's optimal 
     for analysis.
 Depends:
