### 2016-08-24 216997a
Add package-level documentation
