+++ man/plateR.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plateR.R
+\docType{package}
+\name{plateR}
+\alias{plateR}
+\alias{plateR-package}
+\title{Tools to Make it Easy to Work with Microtiter Plate-Shaped Data}
+\description{
+plateR defines a simple, plate-shaped file format for data storage, so it's 
+easy to remember the experimental design. The package provides functions to 
+seamlessly convert between that format and a tidy data frame that's optimal 
+for analysis.\cr\cr You can work with purely plate-shaped data 
+(\code{\link[plateR]{read_plate}} and \code{\link[plateR]{read_plates}}), as 
+well as with a combination of plate-shaped data and tidy data
+(\code{\link[plateR]{add_plate}}). It further allows easy plate-shaped 
+visualization of tidy data (\code{\link[plateR]{view_plate}}).
+}
+
