### 2016-08-24 8fecdb7
Allow use of # in input files. Closes #13
