+++ tests/testthat/test-read_plate.R
@@ -35,6 +35,13 @@ for (i in c(12, 24, 48, 96, 384)) {
          "tbl_df")
    })
    
+   test_that("read_plate allows weird characters", {
+      filename <- paste0(path, "weirdCharacters.csv")
+      
+      result <- read_plate(filename, "wells")
+      expect_that(result$values[1:3], is_identical_to(c("#a", "`abc", "~")))
+   })
+   
    ################################################################################
    context("testing read_plate-calculate_number_of_plates()")
    ################################################################################
