### 2016-08-24 88f212f
Fix bug in view_plate using wrong column name
