### 2016-08-24 e7bef1e
Add tests using plates with single wells only
