### 2016-08-25 7c0b67f
sapply -> vapply mapply -> Map
