### 2016-08-26 #60 @seaaan
# Status update:
Just wanted to let you know where things are since it's been about 10 days. I've been steadily working through @daattali's comments and have only a few left outstanding. After that, I'll do @joolia's. I'll be on vacation for the next two weeks, which honestly means I'll probably spend more time on it, but in case I don't I wanted to let you know where things are. Thank you again for the detailed reviews and also @sckott for organizing -- it's really improving the package.

@daattali see below for responses to almost all of your comments:

# Response to @daattali

## README
- just use README.md
`Point taken, but I prefer to leave it with the .Rmd in case I want to run R code in the README in the future`

- put target audience in README
`added`
https://github.com/seaaan/plateR/commit/784d1701c2c14bd5dad800364b64c64917803c5f

- no Maintainer field in DESCRIPTION
` I didn't include this field based on Hadley's advice to just use Authors@R  but I don't have strong feelings about it and can add a Maintainer field if it's important.`
Hadley's [advice](http://r-pkgs.had.co.nz/description.html)

## Documentation
- add documentation for plateR as a package
`Done!`
https://github.com/seaaan/plateR/commit/216997a

- `file` param in `read_plate` gives the impression you can pass multiple files
`Good point -- this is confusing. Changed it.`
https://github.com/seaaan/plateR/commit/56a7c868

- "blank" row between multiple plate layouts
`This a really good point, thanks. It is meant to be a row of empty cells, so all commas, not strictly a blank row. I've changed the documentation.`
https://github.com/seaaan/plateR/commit/39e9c57

## Functionality
- Why can't you use # in a plateR format file?
`I should have fixed this a long time ago -- now you can!`
https://github.com/seaaan/plateR/commit/8fecdb7

- How is the user meant to generate the input file?
`What I do is make the input file by hand in Excel (gasp). This is intuitive and easy for me. I'm open to ideas for tools that can help with creation, though!`

- For `read_plates()`, put the plate name column as the first column rather than the last
`Good idea. Changed.`
https://github.com/seaaan/plateR/commit/1bdd096

- Why doesn't `well_ids_column` have a default in `add_plate()`?
`My reasoning is that in this case, the data frame passed in to add_plate() didn't come from plateR. Therefore I didn't want to assume anything about the column names and instead force the user to provide it. I thought that for someone not too familiar with the package, getting an error about a missing argument would be less confusing than getting an error about there being no column named "Wells" in their data frame, since if they left it as default they might not have realized that they needed it to match what's in their data frame.`

- Make `data` the first param of `add_plate` for ease of piping.
`Great idea! I love this.`
https://github.com/seaaan/plateR/commit/c8c032c

- It's really helpful to get an error message when reading a plate that tells me exactly in which layout there was a problem
`Thanks!`

- Why can't `view_plate()` figure out the `plate_size`?
`Well, yes and no. If you use, say, the 6 wells in the top left of any plate, the well IDs will be identical across all plates between size 12 and 384. So I could definitely have it guess the plate size, say the smallest size that could fit. However, as a user if I used a 96-well plate but then it displayed as a 12-well plate, I would find this surprising and confusing. Compare:`

`d <- data.frame(Wells = c("A01", "A02", "B01", "B02"), Data = 1:4)`
`plateR::view_plate(d, "Wells", "Data", 12)`
`plateR::view_plate(d, "Wells", "Data", 96)`

`I agree that from a data visualization point of view, showing it as a 12-well plate is superior, but in terms of fidelity to what the user did in the lab, I think this would be confusing.`

## Source code
- `guess_plate_size`: safer to use `vector[-1]` instead of `vector[2:length()]`
`Done`
https://github.com/seaaan/plateR/commit/81f741d

- `get_plate_size_from_number_of_columns`: I don't think the return type is meant to say that it will return TRUE
`Sure isn't!`
https://github.com/seaaan/plateR/commit/7a23b94

- `get_plate_size_from_number_of_columns`: maybe throwing an error makes more sense than returning a string indicating an error
`Yes, that didn't make sense, I just throw the error in the helper instead now`
https://github.com/seaaan/plateR/commit/b0ea5084

- Plate sizes seem random
`Yes, these are the main sizes that are commercially available.`

- I see many uses of mapply and sapply, it might be safer to try to stay away from those functions altogether
`I replaced all instances of vapply with sapply and mapply with Map, which I believe are the safer versions. Is that what you meant? I am happy to learn if there is a better way than this!`
https://github.com/seaaan/plateR/commit/7c0b67f

- Really good use of helper functions
`Thanks!`

- `validate_column_is_in_data()` could take in a vector of columns
`Done! plus added tests`
https://github.com/seaaan/plateR/commit/e1156b7

- I see a lot of dataframe subsettings and very rarely a use of drop = FALSE. Have you verified that all the functions work as intended when the input file has only one well/one row/one column?
`Thanks for the suggestion. I added a bunch of tests for related cases.`
https://github.com/seaaan/plateR/commit/e7bef1e

## Misc
- I wouldn't put LICENSE in buildignore, I'd want CRAN to know about the license
`My understanding is that for certain standard licenses, including GPL-3, you don't need to have a license file and can just list the standard abbreviation in the DESCRIPTION. It's a bit hard to interpret, but it almost sounds like you shouldn't include the license file for standard licenses in the last paragraph in the section of Writing R Extensions. I could definitely be wrong though. I only added the license file because JOSS requires it. This is based on: `
http://r-pkgs.had.co.nz/description.html#license
https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing

- in the YAML of paper.md I would remove the indentation that you have for all the fields
`Done`
https://github.com/seaaan/plateR/commit/2978589

- put depends, imports, suggests in order in DESCRIPTION
`Done`
https://github.com/seaaan/plateR/commit/b5a3082

- add dates to NEWS
`Good idea`
https://github.com/seaaan/plateR/commit/d7888cc

- Add version numbers to cran-comments
`I will definitely do this, but I actually haven't submitted it to CRAN yet, I wanted to go through the code review process first.`

- Awkward second sentence in DESCRIPTION
`Thanks for pointing that out`
https://github.com/seaaan/plateR/commit/54fb599

- plater vs plateR
`I will leave it up to the editors. I guess I prefer plateR for my own convenience in not breaking past code used internally by me and others in the lab but that's not a super good reason.`

## remaining TODOs
- reminder to include rOpenSci footer image after the package gets approved

- explain plateR format elsewhere than just `read_plate` documentation

- Add a separate function that checks a file to see if it is in proper plateR format

- More error checking For example, I shouldn't be allowed to pass `well_ids_column = ""` and I should get a clear error if I pass an invalid string to `file` or a `NULL` value to `well_ids_column`. I also noticed I get a non descriptive error if an input csv file is empty. Don't allow multiple files to `read_plate` or `add_plate` Check that `read_plate` and `add_plate` are only given a single file as input

- `read_plate()` and `read_plates()` are essentially the same: combine them?

- Support multiple columns in `view_plate()`

- I think the only reason you import dplyr is for `bind_row`. If that's indeed the only reason, it might be worth it to just use a base equivalent and remove the dependency on such a heavy package (even though it's safe to assume most people will have it installed). I think something like `do.call(rbind, list_of_df)` will do the job

- read_plate(): it feels a bit hacky to "try adding one in case the file ends in a blank row" -- this method would also succeed is there is an extra row that isn't blank, right? Could you instead explicitly check if the last row is blank and then you know whether or not to remove it?
