### 2016-08-27 #60 @daattali
Great to see so much progress, and that you're really paying attention to every single comment,  Everything you've done looks good, very few remaining things.  I think it's ok if you take some time off during your vacation and come back to this after :)
