### 2016-09-24 e09fbff
Check more rigorously that provided file paths are valid and give a more useful error message if they're not.
