+++ R/read_plate.R
@@ -54,6 +54,8 @@
 #' # Now data are tidy
 #' head(data)
 read_plate <- function(file, well_ids_column = "Wells") {
+   check_file_path(file)
+  
    plate_size <- guess_plate_size(file)
    
    raw_file_list <- get_list_of_plate_layouts(file, plate_size)
