### 2016-09-24 b56fe72
Ensure that the well_ids_column passed in is not null or the empty string.
