### 2016-09-24 81e4d21
remove note that you can't use "#" in the files from the documentation of read_plate
