### 2016-09-24 ce59abc
Ensure good error message if empty csv file is provided
