### 2016-09-24 80c1516
Only allow read_plate and add_plate to accept a file argument of length 1.
