+++ R/read_plate.R
@@ -54,9 +54,10 @@
 #' # Now data are tidy
 #' head(data)
 read_plate <- function(file, well_ids_column = "Wells") {
+   check_that_only_one_file_is_provided(file) 
    check_file_path(file)
-   check_well_ids_column_name(well_ids_column)
    check_that_file_is_non_empty(file)
+   check_well_ids_column_name(well_ids_column)
    
    plate_size <- guess_plate_size(file)
    
