### 2016-09-25 2b90cc9
When calculating the number of plates present in a file, instead of just guessing that the file might end in a blank line, actually check.
