### 2016-09-25 0fc0ca6
Allow display of multiple plate layouts with a single call to view_plate
