### 2016-09-25 71c67ef
Add view_plate example to vignette
