### 2016-09-25 2c6311d
Add version dependency to dplyr (required for dplyr::everything())
