+++ DESCRIPTION
@@ -12,7 +12,7 @@ Depends:
     R (>= 3.1.0)
 Imports:
     utils,
-    dplyr
+    dplyr (>= 0.4.3.9000)
 Suggests:
     testthat,
     knitr,
