### 2016-09-25 9d0b58c
Add example plateR format image to vignette
