### 2016-09-25 8666739
Fix image file path
