### 2016-09-25 e5aed68
Make R CMD CHECK happy about files whose names only differ in case
