### 2016-09-25 17ccd44
Make R CMD CHECK happy by not using NSE.
