+++ R/read_plates.R
@@ -65,7 +65,7 @@ read_plates <- function(files, plate_names = NULL, well_ids_column = "Wells") {
    rownames(result) <- NULL
    
    # make Plate the first column instead of the last column
-   result <- dplyr::select(result, Plate, dplyr::everything())
+   result <- dplyr::select_(result, "Plate", ~ dplyr::everything())
    
    result
 }
