### 2016-09-25 5a274c0
Update travis to run on both R release and R devel, as well as to cache packages
