### 2016-09-25 563a1db
Remove unnecessary sudo: false from Travis config
