+++ .travis.yml
@@ -1,6 +1,5 @@
 language: r
 warnings_are_errors: true
-sudo: false
 cache: packages
 r: 
   - release
