### 2016-09-25 #60 @seaaan
Sorry for the delay in response. I ended up vacationing during my whole vacation :)

@jooolia I'm planning to get to your review tomorrow. Sorry for the wait!

@daattali Thanks again for such a detailed review. Here are responses to the outstanding issues. Let me know what you think:

- More error checking For example, I shouldn't be allowed to pass `well_ids_column = ""` and I should get a clear error if I pass an invalid string to `file` or a `NULL` value to `well_ids_column`. I also noticed I get a non descriptive error if an input csv file is empty.

`Thanks for the detailed examples. I split these up into several commits, shown below:`

* [adds error checking for valid file paths](https://github.com/seaaan/plateR/commit/e09fbff)
* [adds error checking for the names of the well ids column](https://github.com/seaaan/plateR/commit/b56fe72)
* [adds useful error message when empty csv file is provided](https://github.com/seaaan/plateR/commit/ce59abc)
* [makes sure add_plate and read plate only get arguments of length 1](https://github.com/seaaan/plateR/commit/80c1516)

- read_plate(): it feels a bit hacky to "try adding one in case the file ends in a blank row" -- this method would also succeed is there is an extra row that isn't blank, right? Could you instead explicitly check if the last row is blank and then you know whether or not to remove it

`Fixed!`
[Commit](https://github.com/seaaan/plateR/commit/2b90cc9e)

- Support multiple columns in `view_plate()`

`Good idea!`

- [Implemented here](https://github.com/seaaan/plateR/commit/0fc0ca64)
- [Added example to vignette](https://github.com/seaaan/plateR/commit/71c67ef)

- explain plateR format elsewhere than just `read_plate` documentation

`Good suggestion -- I added an image of the format to the vignette.`
[Added example to vignette](https://github.com/seaaan/plateR/commit/9d0b58c)

- `read_plate()` and `read_plates()` are essentially the same: combine them?

`You're definitely right that the only difference between the two functions is that read_plates takes multiple files and a plate names argument while read_plate takes a single file and no plate argument. I did it this way because, as a user, I personally prefer to have more functions with fewer arguments rather than fewer functions with more optional arguments. In this case, it's true that the difference is pretty trivial. I'm inclined to leave them as separate functions but could be persuaded.`

- I think the only reason you import dplyr is for `bind_rows`. If that's indeed the only reason, it might be worth it to just use a base equivalent and remove the dependency on such a heavy package (even though it's safe to assume most people will have it installed). I think something like `do.call(rbind, list_of_df)` will do the job

`I ended up with two calls to dplyr functions: the bind_rows case you noted and then a call to select that I added for reordering columns directly below the call to bind_rows. In essence I used them for convenience and safety:`

`bind_rows nicely handles the case where the data frames don't have all of the same columns, which is possible in read_plates. Unfortunately do.call(rbind, list_of_df) doesn't handle that case. Of course I could write the code to do it, but I guess I trust Hadley to have figured out the possibilities better than I have.`

`I added the call to select for reordering of columns (specifically putting the Plate column first) which is also a pain manually and just so pleasant using select.`

`But I agree that these may not be sufficient reasons to import the package. Let me think about it a bit more and whether you have any other thoughts.`

- Add a separate function that checks a file to see if it is in proper plateR format

`Could you tell me more about how you imagine using this? I can definitely do it, but I don't quite see how it would be helpful. I guess I imagine one would run "check_plateR_format()" and then if it passed, run whichever other plateR function. But all of the other functions already do the checking. I'm definitely open to doing it if there's another aspect I'm missing.`

- reminder to include rOpenSci footer image after the package gets approved

`(still todo)`
