### 2016-09-25 0521fd2
Put back sudo: required in travis
