+++ .travis.yml
@@ -1,6 +1,7 @@
 language: r
 warnings_are_errors: true
 cache: packages
+sudo: required
 r: 
   - release
   - devel
