### 2016-09-25 #60 @daattali
All my suggestions were merely suggestions and points that I wanted to bring up to your attention, so re: "read_plate vs read_plates" and "including dplyr" - I completely support you sticking with the current design if you think that makes the most sense.

check_plateR_format() - seems like it would just be a function I'd expect to exist. As a beginner user of the package, just because your package does require an input format that is non standard, I might want to see if I understand how to create the input files correctly by just getting knowing that my input files passed the validation.  Again, it's your call, but as someone who is going through the readme and trying to learn the package, it was something I really wanted to use in order to prove to myself that I understand the input
