### 2016-09-25 f8b6d97
Add check_plateR_format()
