+++ man/check_plateR_format.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/check_plateR_format.R
+\name{check_plateR_format}
+\alias{check_plateR_format}
+\title{Check whether a file is in plateR format.}
+\usage{
+check_plateR_format(file)
+}
+\arguments{
+\item{file}{The path of the file to check}
+}
+\value{
+Displays a number of messages as it checks the file. Will stop with
+a descriptive error message if the file is not formatted correctly.
+}
+\description{
+Runs the provided file through a number of diagnostics to determine whether
+it is a valid plateR format file and displays information about any 
+deficiencies found.
+}
+\examples{
+file_path <- system.file("extdata", "example-1.csv", package = "plateR")
+
+data <- check_plateR_format(file_path)
+}
+
