+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(add_plate)
+export(check_plateR_format)
 export(read_plate)
 export(read_plates)
 export(view_plate)
