### 2016-09-25 fb6eb43
Add .call = FALSE to stop command
