### 2016-09-25 #60 @seaaan
Thanks, it's very valuable feedback to know that you wished `check_plateR_format()` was there when you were trying to figure out the format. I added the function [here](https://github.com/seaaan/plateR/commit/f8b6d973) and would love your feedback on whether it would be helpful.

Here are some sample calls (they'll work assuming your working directory is the package root):
`check_plateR_format("tests/testthat/testData/96/allWellIds.csv")`
`check_plateR_format("tests/testthat/testData/96/incorrectRowLabels.csv")`
`check_plateR_format("tests/testthat/testData/96/missingBottomRow.csv")`
`check_plateR_format("tests/testthat/testData/96/missingRightColumn.csv")`
`check_plateR_format("tests/testthat/testData/96/missingRowLabels.csv")`
