### 2016-09-25 4556e5e
Add check_plateR_format() to package-level documentation.
