### 2016-09-25 seaaan/plateR#14 @seaaan
Shiny web interface for non-R users
Develop a simple interface where users can submit plateR format files and receive the data back in the shape of their choice.
