### 2016-09-26 #60 @daattali
That's really nice, and great error reporting :+1:
