### 2016-09-26 seaaan/plateR#16 @seaaan
Average replicates
A function or a guide to how to do it
