### 2016-09-27 #60 @jooolia
Hi @seaaan !

Thanks for the thoughtful responses. It really makes reviewing packages even more fun. I have emailed the data and commented on your viz issue. All my comments were just to share my thoughts, I did not have any problems with the package in its current state.

I think it is a very reasonable idea to release your package in its current form and then decide/work on the enhancement ideas.
