### 2016-09-27 4e3a80d
Add test for fixed white space adding bug.
