### 2016-09-27 2f9d57d
Always use `[[` to select a single column from a data frame and get a vector rather than a data frame.
