### 2016-09-27 ce313ce
Make sure add_plate returns an object with all the same classes as passed in.
