### 2016-09-27 #60 @seaaan
Thanks very much @jooolia and @daattali for the time and effort you put into reviewing `plateR`. I know the package is _much_ better now than it was before this process.

I believe that the package is ready to go ahead with onboarding. These are the remaining outstanding issues I can think of:
* Submit to CRAN (it's ready to go, I'm just waiting in case there are any last minute changes that I forgot about)
* Add the rOpenSci image to the README (just waiting for official approval)
* Resolve the plater/plateR naming issue (I would prefer to leave it as plateR, but if lower-case is a strong preference of rOpenSci, I'm okay with changing it. What do you think @sckott?)
* ???

Thanks again all, this was very valuable to me!
