### 2016-09-27 7ba34a8
Minor rewordings of description and JOSS paper
