### 2016-09-27 04259cd
Version bump to 1.0.0 and updated details for CRAN submission
