+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: plateR
-Version: 0.2.1
 Title: Tools to Make it Easy to Work with Data from Microtiter Plates
+Version: 1.0.0
 Authors@R: person("Sean", "Hughes", email = "smhughes@uw.edu",
     role = c("aut", "cre"))
 Description: Tools for interacting with data from experiments done in microtiter
