### 2016-09-29 #60 @seaaan
Woohoo! I'll do this over the next few days, once it gets approved and put
on CRAN. A couple questions for you @sckott:
- If I plan to work on other outstanding issues in the future, should I do
that in a fork on my own github account or in the rOpenSci repo? Related,
are you saying that there shouldn't be any open issues in the rOpenSci one
or just that I can remove them if I want to?
- Does rOpenSci go in the DESCRIPTION?

Thanks!

On Wed, Sep 28, 2016 at 1:59 PM, Scott Chamberlain <notifications@github.com
> wrote:

> thanks for the reviews @daattali <https://github.com/daattali> and
> @jooolia <https://github.com/jooolia> 🚀 ✏️ 📋
>
> @seaaan <https://github.com/seaaan> Approved!
>
> Approved! Here are the remaining steps:
>
>    - Close any other outstanding issues if you'd like.
>    - Add the rOpenSci footer to the README
>    - Transfer the repo to ropenscilabs - I've created a team in that org.
>    account, and you should have received an email about, after you accept the
>    invitation, go to Settings->Transfer Ownership
>    - After transferring to ropenscilabs, travis builds should work
>    automatically under the new github account. but you do need to update the
>    link in the badge in your readme
>    - Package name issue: I'd encourage you to lowercase the package name
>    since I think the most important aspect of the name is that people will
>    type it over and over again, and surely many will forget that the "r"
>    should be capitalized. The case won't make a difference for google indexing
>    it either, so that doesn't matter for find-ability of usage of your
>    package. Decision is up to you.
>    - Update all links to github repo to replace seaaan with ropenscilabs
>    - I noticed one thing to deal with after running goodpractice::gp() at
>    some point (no need to do prior to transfering though):
>       - Avoid using 1:nrow(...), use seq_len() or seq_along instead
>
