### 2016-09-29 #60 @sckott
> If I plan to work on other outstanding issues in the future, should I do
that in a fork on my own github account or in the rOpenSci repo?

Do changes in the ropenscilabs version. Treat it as you did when it was under your account.

> Related, are you saying that there shouldn't be any open issues in the rOpenSci one
or just that I can remove them if I want to?

No, just check to see if there's any issues you hadn't yet dealt with from these reviews

> Does rOpenSci go in the DESCRIPTION?

Just in the `URL` and `BugReports` https://github.com/seaaan/plateR/blob/master/DESCRIPTION#L20-L21 fields since the url will change to `.../ropenscilabs/...` instead of your user name
