### 2016-09-30 c11ac4c
Updated CRAN comments before submission. [ci ignore]
