+++ cran-comments.md
@@ -1,21 +1,28 @@
+# Initial submission, version 1.0.0, 27 Sep 2016
+
 ## Test environments
-* local Windows 7, R 3.2.3
-* local Ubuntu 14.04, R 3.2.2
-* ubuntu 12.04 (on travis-ci), R 3.3.1
-* win-builder (devel (2016-07-25 r70972) and release (3.3.1 (2016-06-21))
+* ubuntu 12.04 on travis-ci:  devel   r71384 (2016-09-27) 
+                              release 3.3.1  (2016-06-21)
+* win-builder:                devel   r71370 (2016-09-26) 
+                              release 3.3.1  (2016-06-21)
+* local Windows 7:            release 3.2.3  (2015-12-10)
+* local Ubuntu 14.04          release 3.3.1  (2016-09-27)
 
 ## R CMD check results
 There were no ERRORs or WARNINGs. 
 
-There was one NOTE, indicating that this is a new submission. 
+There was one NOTE:
 
    * checking CRAN incoming feasibility ... NOTE
    Maintainer: 'Sean Hughes <smhughes@uw.edu>'
    New submission
 
    Possibly mis-spelled words in DESCRIPTION:
-     Microtiter (2:43)
-     microtiter (6:61)
-     
+     Microtiter (2:53)
+     microtiter (6:71)
+   
+   This is a valid spelling of the word (eg         
+      https://en.wikipedia.org/wiki/Microtiter_plate).
+      
 ## Downstream dependencies
 There are currently no downstream dependencies for this package.
