+++ inst/doc/plater-basics.R
@@ -3,15 +3,15 @@
 # output:
 #   md_document:
 #     variant: markdown_github
-# then switch back to output: rmarkdown::html_vignette. Just open plateR-basics.md and resave to update time stamp
+# then switch back to output: rmarkdown::html_vignette. Just open plater-basics.md and resave to update time stamp
 
-library(plateR)
+library(plater)
 
 # print results of code using #>
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 ## ------------------------------------------------------------------------
-file_path <- system.file("extdata", "example-1.csv", package = "plateR")
+file_path <- system.file("extdata", "example-1.csv", package = "plater")
    
 data <- read_plate(
       file = file_path,             # full path to the .csv file
@@ -22,14 +22,14 @@ str(data)
 head(data)
 
 ## ------------------------------------------------------------------------
-file2A <- system.file("extdata", "example-2-part-A.csv", package = "plateR")
+file2A <- system.file("extdata", "example-2-part-A.csv", package = "plater")
 data2 <- read.csv(file2A)
 
 str(data2)
 
 head(data2)
 
-meta <- system.file("extdata", "example-2-part-B.csv", package = "plateR")
+meta <- system.file("extdata", "example-2-part-B.csv", package = "plater")
 data2 <- add_plate(
       data = data2,               # data frame to add to 
       file = meta,                # full path to the .csv file
@@ -42,10 +42,10 @@ head(data2)
 
 ## ------------------------------------------------------------------------
 # same file as above
-file1 <- system.file("extdata", "example-1.csv", package = "plateR")
+file1 <- system.file("extdata", "example-1.csv", package = "plater")
 
 # new file
-file2 <- system.file("extdata", "more-bacteria.csv", package = "plateR")
+file2 <- system.file("extdata", "more-bacteria.csv", package = "plater")
 
 data <- read_plates(
    files = c(file1, file2),
