+++ R/plater.R
@@ -1,0 +1,16 @@
+#' Tools to Make it Easy to Work with Microtiter Plate-Shaped Data
+#' 
+#' plater defines a simple, plate-shaped file format for data storage, so it's 
+#' easy to remember the experimental design. The package provides functions to 
+#' seamlessly convert between that format and a tidy data frame that's optimal 
+#' for analysis. \code{\link[plater]{check_plater_format}} is provided to help 
+#' you manage plate-shaped files. \cr\cr You can work with purely plate-shaped 
+#' data (\code{\link[plater]{read_plate}} and 
+#' \code{\link[plater]{read_plates}}), as well as with a combination of 
+#' plate-shaped data and tidy data (\code{\link[plater]{add_plate}}). It further
+#' allows easy plate-shaped visualization of tidy data 
+#' (\code{\link[plater]{view_plate}}). 
+#' 
+#' @docType package
+#' @name plater
+NULL
