### 2016-10-04 4f6dcc4
Fix missed name changes.
