### 2016-10-04 2e1bc0a
use seq_len instead of 1:length
