+++ R/convert_plate_to_column.R
@@ -33,7 +33,7 @@ convert_plate_to_column <- function(plate, plate_size) {
    cols <- ncol(plate)
    
    # convert to vector
-   plate <- unlist(lapply(1:nrow(plate), function(i) unname(plate[i, ])))
+   plate <- unlist(lapply(seq_len(nrow(plate)), function(i) unname(plate[i, ])))
    # as.vector(t(plate)) is simpler, but t() calls as.matrix() and when you have 
    # a plate layout with one numeric column and one character column, all are
    # converted to character, but numeric goes to character via format() which 
