### 2016-10-04 2dd6832
update cran-comments
