### 2016-10-04 #60 @sckott
awesome, thanks @seaaan - closing this issue now. let us know if you have any further questions.
