### 2016-10-06 #60 @seaaan
On [CRAN](https://cran.r-project.org/web/packages/plater/index.html)! No binaries yet.

@sckott, What do I need to do for the JOSS paper?
