### 2016-10-06 #60 @noamross
For JOSS:

-  Archive your package in a long-term repository with a DOI.  We recommend Zenodo.  Here are instructions: https://guides.github.com/activities/citable-code/
-  Check off the DOI box and add the DOI at the top of this review
-  Submit at http://joss.theoj.org/papers/new
-  We will alert the JOSS editor that this is an rOpenSci package, so they can make an expedited decision based on this thread rather than assign new reviewers.
