### 2016-10-19 #60 @noamross
Hi @seaan, just checking in, should we expect this to be submitted to JOSS?
