### 2016-10-19 #60 @sckott
@seaaan it looks like Zenodo is an approved 3rd party application for that org account. What happens when you try to give Zenodo access?
